Правка правил удаления.

+++ app/Http/Controllers/CategoriesController.php
@@ -5,30 +5,42 @@ namespace App\Http\Controllers;
 use App\Categories;
 use App\Post;
 use Illuminate\Http\Request;
+use App\Utilities\СheckWhoUpdated;
 
 class CategoriesController extends Controller
 {
 //  Метод вывода всех Категорий.
 
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function index()
     {
         $data = Categories::all();
         $data = Categories::getAmountPosts($data);
-        return response()->json($data,200);
+        return response()->json($data, 200);
     }
 
 //  Метод вывода конкретной Категории.
 
+    /**
+     * @param Categories $category
+     * @return \Illuminate\Http\JsonResponse
+     */
     public function show(Categories $category)
     {
         $data = $category;
-        $data['posts'] = Post::All()->where('category_id',$category->id)->count();
-        return response()->json($data,200);
+        $data['posts'] = Post::All()->where('category_id', $category->id)->count();
+        return response()->json($data, 200);
     }
 
 //  Метод отправки содержимого форм в БД.
 
-    public function store()
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function store(Request $request)
     {
         request()->validate([
             'name' => 'required',
@@ -45,20 +57,31 @@ class CategoriesController extends Controller
 
 //  Метод сохранения изменений в Категории.
 
-    public function update(Categories $category) {
+    /**
+     * @param Categories $category
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function update(Categories $category)
+    {
         //todo Добавить валидацию
-        $category->update(request()->all());
-
-        return response()->json($category,200);
+        if (СheckWhoUpdated::check($category['user_id'])) {
+            $category->update(request()->all());
+            return response()->json($category, 200);
+        } else {
+            return response()->json(['Error' => 'You don\' have rule'], 403);
+        }
     }
 
 //  Метод Удаления Категории.
 
-    public function destroy($id)
+    public function destroy(Categories $category)
     {
-        Categories::findOrFail($id)->delete();
-
-        return response()->noContent(204);
+        if (СheckWhoUpdated::check($category['user_id'])) {
+            Categories::findOrFail($id)->delete();
+            return response()->noContent(204);
+        } else {
+            return response()->json(['Error' => 'You don\' have rule'], 403);
+        }
     }
 }
